add gate permision for divisi controller

+++ app/Http/Controllers/DivisiController.php
@@ -7,6 +7,7 @@ use App\Models\Divisi;
 use App\Http\Requests\StoreDivisiRequest;
 use App\Http\Requests\UpdateDivisiRequest;
 use Illuminate\Http\RedirectResponse;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\View\View;
 
 class DivisiController extends Controller
@@ -20,39 +21,39 @@ class DivisiController extends Controller
 
     public function index(): View
     {
+        abort_if(!Gate::allows('view division'), 403);
+
         $divisis = $this->repository->getDivisi();
         return view('divisi.index', compact('divisis'));
     }
 
-    public function create()
-    {
-        //
-    }
-
     public function store(StoreDivisiRequest $request): RedirectResponse
     {
+        abort_if(!Gate::allows('create division'), 403);
+
         $this->repository->store($request->validated());
         return redirect()->route('divisi.index')->with('toastr-success', 'Divisi Successfully Created');
     }
 
-    public function show(Divisi $divisi)
-    {
-        //
-    }
-
     public function edit(Divisi $divisi): View
     {
+        abort_if(!Gate::allows('edit division'), 403);
+
         return view('divisi.edit', compact('divisi'));
     }
 
     public function update(UpdateDivisiRequest $request, Divisi $divisi): RedirectResponse
     {
+        abort_if(!Gate::allows('edit division'), 403);
+
         $this->repository->update($request->validated(), $divisi);
         return redirect()->route('divisi.index')->with('toastr-success', 'Divisi Successfully Updated');
     }
 
     public function destroy(Divisi $divisi): RedirectResponse
     {
+        abort_if(!Gate::allows('delete division'), 403);
+
         $this->repository->delete($divisi);
         return redirect()->route('divisi.index')->with('toastr-success', 'Divisi Successfully Deleted');
     }
